refactor: resolve PR review comments on recency and MCP changes
- Move recency functions to utils.recency to eliminate circular imports
- Update MCP methods to return Pydantic objects instead of JSON strings
- Move imports to top of files now that circular dependencies are resolved
